refactor(prompt): move prompt_* function implementation to eval/buffer.c
